Making tests use test-specific abbreviation helper-functions.
